requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.5.0
python-dotenv>=1.0.0
fake-useragent>=1.1.1
logging>=0.5.1.2